+++ parser/sql/engine/dialect/hive/src/main/java/org/apache/shardingsphere/sql/parser/engine/hive/visitor/statement/type/HiveDDLStatementVisitor.java
@@ -66,6 +66,7 @@ import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.constrain
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.index.IndexNameSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.index.IndexSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.table.CherryPickDefinitionSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.table.TableRollbackDefinitionSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.column.ColumnSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.expr.simple.LiteralExpressionSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.generic.DataTypeSegment;
@@ -96,6 +97,7 @@ import org.apache.shardingsphere.sql.parser.statement.core.statement.type.dml.Se
 import org.apache.shardingsphere.sql.parser.statement.core.value.collection.CollectionValue;
 import org.apache.shardingsphere.sql.parser.statement.core.value.identifier.IdentifierValue;
 import org.apache.shardingsphere.sql.parser.statement.core.value.literal.impl.NumberLiteralValue;
+import org.apache.shardingsphere.sql.parser.statement.core.value.literal.impl.StringLiteralValue;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -157,6 +159,22 @@ public final class HiveDDLStatementVisitor extends HiveStatementVisitor implemen
             LiteralExpressionSegment snapshotId = new LiteralExpressionSegment(numberStartIndex, numberStopIndex,
                     new NumberLiteralValue(ctx.cherryPickClause().NUMBER_().getText()).getValue());
             result.cherryPickDefinition(new CherryPickDefinitionSegment(ctx.cherryPickClause().EXECUTE().getSymbol().getStartIndex(), numberStopIndex, snapshotId));
+        }
+        if (null != ctx.tableRollback()) {
+            int startIndex = ctx.tableRollback().EXECUTE().getSymbol().getStartIndex();
+            if (null != ctx.tableRollback().string_()) {
+                int stringStartIndex = ctx.tableRollback().string_().getStart().getStartIndex();
+                int stringStopIndex = ctx.tableRollback().string_().getStop().getStopIndex();
+                LiteralExpressionSegment timestamp = new LiteralExpressionSegment(stringStartIndex, stringStopIndex,
+                        new StringLiteralValue(ctx.tableRollback().string_().getText()).getValue());
+                result.tableRollbackDefinition(new TableRollbackDefinitionSegment(startIndex, stringStopIndex, timestamp, null));
+            } else {
+                int numberStartIndex = ctx.tableRollback().NUMBER_().getSymbol().getStartIndex();
+                int numberStopIndex = ctx.tableRollback().NUMBER_().getSymbol().getStopIndex();
+                LiteralExpressionSegment snapshotId = new LiteralExpressionSegment(numberStartIndex, numberStopIndex,
+                        new NumberLiteralValue(ctx.tableRollback().NUMBER_().getText()).getValue());
+                result.tableRollbackDefinition(new TableRollbackDefinitionSegment(startIndex, numberStopIndex, null, snapshotId));
+            }
         }
         return result.build();
     }